ESOLA-133 Add quorum settings to support only 2 nodes remaining.
Added the two quorum settings "two_node" and "auto_tie_breaker" to allow
50% of the nodes are offline without killing the cluster.

Note that this change doesn't take into consideration of which 50% of the
nodes go down. The cluster may still be up even though all nodes in a
HA pair are killed. This scenario will be handled separately.